backend/requirements.txt
asyncpg~=0.30.0
fastapi~=0.111.0
ffmpeg-python~=0.2.0
moviepy~=1.0.3
numpy>=2.0.0
openai~=1.12.0
pydantic~=2.7.0
pydantic-settings~=2.2.0
PyJWT~=2.8.0
redis~=5.0.0
rq~=1.15.0
sentence-transformers~=2.5.0
slowapi~=0.1.9
supabase~=2.3.0
uvicorn[standard]~=0.29.0
python-multipart~=0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
celery==5.3.6
sqlmodel==0.0.14
pytest==8.0.0
